pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "crewai>=1.13.0",
    "crewai-tools>=1.13.0",
    "fastapi>=0.136.1",
    "openai>=2.36.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "python-socketio>=5.16.1",
    "uvicorn[standard]>=0.46.0",
]
